requirements.txt
torch
torchvision
numpy
pandas
matplotlib
seaborn
opencv-python
scikit-learn
tqdm
Pillow
tensorboard
streamlit
wandb
grad-cam
requests
gdown
sentence-transformers